+++ lib/customize.dart
@@ -6,5 +6,8 @@ export 'src/fx_extensions.dart';
 ///Utilities Properties
 export 'src/fx_utilities.dart';
 
+///Faker Methods
+export 'src/core/faker/fx_faker.dart';
+
 ///Custom Widgets
 export 'src/fx_widgets.dart';