+++ lib/src/core/faker/fx_faker.dart
@@ -1,0 +1,3 @@
+import 'package:customize/src/core/faker/generator/lorem.dart';
+
+class Faker with Lorem {}